Remove parallelism in the HoodieLocalContextEngine
The HoodieLocalContextEngine is used when running within Hive. It uses parallel() and JVM threads to parallelize operations.
This is a problem when using it in the context of a doAs operation. In this scenario some of the executions will be delegated
to the JVM threads which don't have the doAs context set